Apply clang tidy to data_provider and dbsync

+++ src/common/data_provider/src/hardware/architectureDependantX86_64Mac.cpp
@@ -1,14 +1,3 @@
-/*
- * Wazuh SYSINFO
- * Copyright (C) 2015, Wazuh Inc.
- * May 4, 2023.
- *
- * This program is free software; you can redistribute it
- * and/or modify it under the terms of the GNU General Public
- * License (version 2) as published by the FSF - Free Software
- * Foundation.
- */
-
 #include "hardwareWrapperImplMac.h"
 
 int getMhz(IOsPrimitivesMac* osPrimitives)
@@ -23,5 +12,5 @@ int getMhz(IOsPrimitivesMac* osPrimitives)
         throw std::system_error {ret, std::system_category(), "Error reading cpu frequency."};
     }
 
-    return cpuHz / MHz;
+    return static_cast<int>(cpuHz / MHz);
 }
